New class that manages BrushProperties
